+++ HISTORY.md
@@ -1,5 +1,9 @@
 History of features added to LiChess Tools in time (reversed order)
 
+14 Apr
+  - Resize Explorer
+  - Cut result, Search by index, Search invalid games, persistence, send to in PGN Editor
+
 08 Apr
   - Highlight checks